requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
